refactor: rename Web3Provider to EthereumProvider
