+++ packages/ethereum/src/index.ts
@@ -4,4 +4,3 @@ export { chains } from '@ant-design/web3-common';
 
 export * from './ethereum-provider';
 export * from './universal-provider';
-export * from './web3-provider';
