--- packages/ethereum/src/web3-provider.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Web3ConfigProvider } from '@ant-design/web3-common';
-import { EthereumProvider } from './ethereum-provider';
-import { UniversalProvider } from './universal-provider';
-
-export interface Web3ProviderProps {
-  provider: EthereumProvider;
-  children?: React.ReactNode;
-}
-
-export const Web3Provider: React.FC<Web3ProviderProps> = (props) => {
-  const provider = new UniversalProvider(props.provider);
-  return <Web3ConfigProvider provider={provider}>{props.children}</Web3ConfigProvider>;
-};
