+++ packages/web3/package.json
@@ -25,6 +25,7 @@
     "classnames": "^2.3.2"
   },
   "devDependencies": {
+    "@ant-design/web3-ethereum": "workspace:*",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
     "father": "^4.1.9",
